Build: Fix JaCoCo XML report paths
Goal ``report-aggregate'' is not supposed to be configured at each
module. In a multi-module project, you need to ensure a single module
calls ``report-aggregate''. Can parent module be that single module?
No, it cannot. Because it will be executed first, while aggregation
needs to be executed last. A hacked solution is to use a separated
module, e.g. ``jacoco-report-module'', to handle it. Also, this
module needs to have explicit dependencies on all other modules, so
that it can be executed last.

Refs:

- jacoco-maven-plugin: aggregated report empty with separate
  aggregation and inheritance pom
  <https://groups.google.com/forum/#!topic/jacoco/FpdLbxsXSTY>